requirements.txt
anyio==3.4.0
asgiref==3.4.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2020.6.20
charset-normalizer==2.0.9
click==8.0.3
fastapi==0.70.1
firebase-admin==5.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
msgpack==1.0.3
oauthlib==3.1.1
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.16.0